Configure Webpack with code splitting on client and not on server
Okay, so... if we’re gonna have multiple react_on_rails registered
components on the same page, we need to register them in the same
registry. Non-code-split bundles get in the way of this because each new
bundle overwrites the global registry. We must include ReactOnRails in
only one vendor bundle.

However, we can’t have code splitting enabled for server rendering,
since that needs one clean entry for all components. The vendor bundle
won’t be loaded and we’ll get errors all over the place.

The solution is to have two webpack configurations and run two watch
servers... which would be way easier if we weren’t using webpacker.
Convention over configuration breaks down when you’re doing something so
unconventional. When NODE_ENV is set to server by the second webpack
entry in the procfile, we only compile server.entry.js, and we disable
code splitting and hashed naming (since each webpack process overwrites
the manifest, and we don’t need cachebusting if we never send this
entry to the client anyway).